refactor: Seperate out consume and produce interfaces
