+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/EncodedOffsetPair.java
@@ -4,7 +4,8 @@ package io.confluent.parallelconsumer;
  * Copyright (C) 2020 Confluent, Inc.
  */
 
-import io.confluent.parallelconsumer.ParallelConsumerImpl.Tuple;
+
+import io.confluent.parallelconsumer.ParallelConsumer.Tuple;
 import lombok.Getter;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
