+++ parallel-consumer-core/src/test/java/io/confluent/parallelconsumer/StreamingParallelConsumerImplTest.java
@@ -4,7 +4,7 @@ package io.confluent.parallelconsumer;
  * Copyright (C) 2020 Confluent, Inc.
  */
 
-import io.confluent.parallelconsumer.ParallelConsumerImpl.ConsumeProduceResult;
+import io.confluent.parallelconsumer.ParallelConsumeThenProduce.ConsumeProduceResult;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.producer.ProducerRecord;
